Lets strip out the rpcplus code
